Add ability to change the 'activeClass' name
Previously, the active class name was hard coded to ‘active’. Now it
will pull the child `{{link-to}}` `activeClass` name by default. Or,
users can specify what it will be by defining the `activeClass` on the
`{{active-link}}`. Also added tests for this and updated the README
(which has been expanded to show more examples).