Add support for the disabled state
As a negative to active, add a disabled state that adds a class when
ALL child links are also disabled. Similarly, the class name draws from
the child `{{link-to}}` but can be overridden. Users can always opt-out
by defining `disabledClass=false`. Also added tests for this and
updated the README.